Remove generic `is` impl and only support `is` when one of the arg is None. And, move test into test_optional.py
